+++ CHANGE_LOG.md
@@ -1,2 +1,1 @@
-* Fix putting fish into a fish plaque will spawn fish at 0, 0, closes #54
-* Update uk_ua.json (#52) - unroman
+* Improved Islehopper spawning, closes #58